+++ actix-multipart/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.6.2
+
 - Add testing utilities under new module `test`.
 - Minimum supported Rust version (MSRV) is now 1.72.