+++ actix-test/CHANGES.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 0.1.4
+
+- Add `TestServerConfig::rustls_0_23()` method for Rustls v0.23 support behind new `rustls-0_23` crate feature.
+- Add `TestServerConfig::disable_redirects()` method.
+- Various types from `awc`, such as `ClientRequest` and `ClientResponse`, are now re-exported.
 - Minimum supported Rust version (MSRV) is now 1.72.
 - TestServer can now be told to listen on an interface other than `localhost`, via the `TestServerConfig::listen_address()` method.